backend/pyproject.toml
[project]
name = "polyphony-ledger-backend"
version = "0.1.0"
description = "Polyphony Ledger backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.116.1",
  "uvicorn[standard]==0.35.0",
  "asyncpg==0.30.0",
  "confluent-kafka==2.11.1",
  "clickhouse-connect==0.8.18",
  "prometheus-client==0.23.1",
  "pydantic==2.11.7",
  "pydantic-settings==2.10.1",
  "python-dotenv==1.1.1",
  "orjson==3.11.2",
  "httpx==0.28.1",
  "opentelemetry-api==1.37.0",
  "opentelemetry-sdk==1.37.0",
  "opentelemetry-exporter-otlp-proto-grpc==1.37.0",
  "opentelemetry-instrumentation-fastapi==0.58b0",
  "protobuf==6.32.0",
  "grpcio-tools==1.74.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
